Moves AnnotationMacros.hpp up one directory
Since users are more likely to directly include this one.
