+++ src/axom/core/tests/utils_annotations.cpp
@@ -8,7 +8,7 @@
 #include "axom/config.hpp"
 #include "axom/core/Macros.hpp"
 #include "axom/core/utilities/Annotations.hpp"
-#include "axom/core/utilities/AnnotationMacros.hpp"
+#include "axom/core/AnnotationMacros.hpp"
 
 #include "axom/fmt.hpp"
 #include "axom/CLI11.hpp"
